src/main/java/br/com/ghwh/issuewh/config/security/WebSecurityConfig.java
package br.com.ghwh.issuewh.config.security;

import org.springframework.context.ApplicationContextAware;
import org.springframework.context.annotation.Configuration;
import org.springframework.security.config.annotation.web.builders.HttpSecurity;
import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;

@Configuration
@EnableWebSecurity
public class WebSecurityConfig extends WebSecurityConfigurerAdapter implements ApplicationContextAware {
    
    @Override
    public void configure(final HttpSecurity http) throws Exception {
        http
            .csrf()
                .disable();
    }
}
